Enable class FDR by default instead of Global FDR.
