+++ pypgatk/commands/peptide_class_fdr.py
@@ -5,19 +5,12 @@ from pypgatk.commands.utils import print_help
 import pkgutil
 
 from pypgatk.proteomics.openms import OpenmsDataService
-from pypgatk.toolbox.general import read_yaml_from_text, read_yaml_from_file
+from pypgatk.toolbox.general import read_yaml_from_text, read_yaml_from_file, parse_peptide_classes, \
+  parse_peptide_groups
+
 default_config_text = pkgutil.get_data(__name__, "../config/openms_analysis.yaml").decode()
 
 log = logging.getLogger(__name__)
-
-def parse_peptide_groups(peptide_groups_prefix):
-  peptide_groups = {}
-  for group in peptide_groups_prefix.split(";"):
-    lt = group.split(":")
-    class_group = lt[0].replace("{","")
-    classes = [x.replace("[","").replace("]","") for x in lt[1].split(",")]
-    peptide_groups[class_group] = classes
-  return peptide_groups
 
 @click.command('peptide-class-fdr', short_help="Command to compute the Peptide class FDR")
 @click.option('-c', '--config_file', help='Configuration to perform Peptide Class FDR')
@@ -28,10 +21,11 @@ def parse_peptide_groups(peptide_groups_prefix):
 @click.option('--psm-pep-class-fdr-cutoff', help="PSM class peptide FDR cutoff or threshold", default=0.01)
 @click.option('--peptide_groups_prefix', help="Peptide class "
               "groups e.g. \"{non_canonical:[altorf,pseudo,ncRNA];mutations:[COSMIC,cbiomut];variants:[var_mut,var_rs]}\"")
+@click.option('--peptide_classes_prefix', help='Peptides classes e.g. \"altorf,pseudo,ncRNA,COSMIC,cbiomut,var_mut,var_rs\"')
 @click.option("--enable_class_fdr", help="Enable Class-FDR over Global PSM FDR (default true)", default = True)
 @click.pass_context
 def peptide_class_fdr(ctx, config_file, input_idxml, output_idxml, min_peptide_length, psm_pep_fdr_cutoff, psm_pep_class_fdr_cutoff,
-                      peptide_groups_prefix, enable_class_fdr):
+                      peptide_groups_prefix, peptide_classes_prefix, enable_class_fdr):
 
   if config_file is None:
     config_data = read_yaml_from_text(default_config_text)
@@ -51,6 +45,12 @@ def peptide_class_fdr(ctx, config_file, input_idxml, output_idxml, min_peptide_l
   if psm_pep_fdr_cutoff is not None:
     pipeline_arguments[OpenmsDataService.CONFIG_PEPTIDE_FDR_CUTOFF] = psm_pep_fdr_cutoff
 
+  if peptide_classes_prefix is not None and peptide_groups_prefix is not None:
+    raise ValueError("The tool can't be use to compute class groups and classes FDR at the same time")
+
+  if peptide_classes_prefix is not None:
+    data = parse_peptide_classes(peptide_classes_prefix)
+    pipeline_arguments[OpenmsDataService.CONFIG_PEPTIDE_GROUP_PREFIX] = data
 
   if psm_pep_class_fdr_cutoff is not None:
     pipeline_arguments[OpenmsDataService.CONFIG_PEPTIDE_CLASS_FDR_CUTOFF] = psm_pep_class_fdr_cutoff
